Move TfTypesValueFromGoTypeValue to testutils
