+++ ec/ecresource/deploymentresource/elasticsearch/v2/node_types_validator_test.go
@@ -23,6 +23,7 @@ import (
 
 	deploymentv2 "github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/deployment/v2"
 	v2 "github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/elasticsearch/v2"
+	"github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/testutil"
 	"github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/utils"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
@@ -66,7 +67,7 @@ func TestNodeTypesValidator_ValidateString(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			v := v2.VersionSupportsNodeTypes()
-			config := tftypesValueFromGoTypeValue(t, &deploymentv2.Deployment{
+			config := testutil.TfTypesValueFromGoTypeValue(t, &deploymentv2.Deployment{
 				Version: tt.version,
 			}, deploymentv2.DeploymentSchema().Type())
 			resp := validator.StringResponse{}
